+++ Detour/Include/DetourMath.h
@@ -12,13 +12,13 @@ Members in this module are wrappers around the standard math library
 // if cmath is included before math.h.
 #include <cmath>
 
-inline float dtMathFabsf(float x) { return fabsf(x); }
-inline float dtMathSqrtf(float x) { return sqrtf(x); }
-inline float dtMathFloorf(float x) { return floorf(x); }
-inline float dtMathCeilf(float x) { return ceilf(x); }
-inline float dtMathCosf(float x) { return cosf(x); }
-inline float dtMathSinf(float x) { return sinf(x); }
-inline float dtMathAtan2f(float y, float x) { return atan2f(y, x); }
-inline bool dtMathIsfinite(float x) { return std::isfinite(x); }
+inline double dtMathFabsf(double x) { return fabs(x); }
+inline double dtMathSqrtf(double x) { return sqrt(x); }
+inline double dtMathFloorf(double x) { return floor(x); }
+inline double dtMathCeilf(double x) { return ceil(x); }
+inline double dtMathCosf(double x) { return cos(x); }
+inline double dtMathSinf(double x) { return sin(x); }
+inline double dtMathAtan2f(double y, double x) { return atan2(y, x); }
+inline bool dtMathIsfinite(double x) { return std::isfinite(x); }
 
 #endif
